WL#11563 Post-push fix to test ndb.ndbinfo
The testing strategy for this worklog revolves around using assert_grep.inc
to test for specific messages in the MySQL server error log. When the schema
should be upgraded, we look for the "Installing ndbinfo schema" message;
when it should not upgraded, we look for the message "Installed ndbinfo
schema is current. Not upgrading."

The test for the "not upgrading" scenario was in ndbinfo.test and
tried to avoid restarting the server. This is unreliable. Move the
test code to a new test case called ndbinfo_no_upgrade, and force a
server restart before testing.

Change-Id: I8b11612dfb491aab38efed1d129fae5cca466094