WL#13773: Increase default group_replication_member_expel_timeout
Change the default value of variable group_replication_member_expel_timeout from
0 to 5 seconds.

The default value of group_replication_member_expel_timeout is currently set to
0 seconds, which means that a member which is suspected of leaving the group is
expelled immediately. Since there is a 5 second waiting period  until creating
the suspicion that the member has left the group, the member is actually
expelled 5 seconds after becoming unreachable. If the expelled member is the
primary a new primary is elected automatically.

Set the default value group_replication_member_expel_timeout to 5 seconds, in
order to decrease the likelihood of unnecessary expulsions and primary failovers
on slower networks or in the presence of transient network failures, following
feedback gathered from customers.

The new value for the default implies that the member will be expelled 10
seconds after becoming unreachable.

RB: 23626
Reviewed-by : Nuno Carvalho <nuno.carvalho@oracle.com>
Reviewed-by : Parveez Baig <parveez.baig@oracle.com>