BUG#25660161 - GR MEMBER STATUS IS WRONG
Description:
When I network partition a node, it remains ONLINE, nothing is indicating
there's a problem.

Analysis:

GCS/XCom has a full mesh of connections between all nodes. A connects to B and
B connects back to A.

The issue is visible when we cut out one side of the connection. If we cut out
B with, for instance, a firewall,we have the A->B connection silently dead,
but we have the B->A connection alive.Since we only do monitoring on one half
of the connection (the incoming messages), we will consider that B is alive,
although we can't contact it. In the same way, B will consider that A is dead,
since it does not receive any message from it.

Fix:

We must be able to break the outgoing connection if we detect that something
is wrong. That can be done if we start receiving pings from a node, in which
we consider that it is alive. After some pings, we just kill the outgoing
connection, thus creating a consistent state.

RB: 23671