Bug#30199859 EXTRA OUTPUT WHEN COMPILING FUZZER TESTS
Problem
-------

Compiling mysql with ninja usually gets no extra output.

When building with clang and libfuzzer is available and with

  cmake -DWITH_UNIT_TESTS=1

extra output is generated when building the routerfuzz targets.

Change
------

The INFO lines can be hidden by adding the command-line params
'-verbosity=0', but the 'MERGE-INNER' and 'MERGE-OUTER' lines
can't be. Therefore, redirecting all output to /dev/null.

- redirect output of the merge step to /dev/null

RB: 23770