BUG#30846046: BINLOG COMPRESSION DOES NOT RESET CACHE EVENT COUNTER BEFORE COMPRESSION
During a transaction execution, its changes are captured as events
in the binlog cache. For instance:
  INSERT INTO t1 VALUES(1);
will generate 5 events:
  1) GTID;
  2) Query_log_event (BEGIN);
  3) Table_map;
  4) Write_rows;
  5) Query_log_event (COMMIT).
When binlog compression is enable,
--binlog-transaction-compression=ON, the content of the binlog cache
is replaced by only two events:
  1) GTID;
  2) Transaction_Payload.
The transaction size is computed as the sum of:
  * cache size;
  * binlog checksum size * number of events on the cache;
  * GTID event size + its checksum.
This transaction size is saved on the GTID event.

When binlog compression replaced the content of the cache, the
number of events contained on it was not reset, that resulted on the
above example that 6 instead of 2 events were considered.
On its turn, that was making the transaction length considering
extra bytes, in this case 4 * 4 (checksum size).
At the end, we had a transaction with X bytes on the binary log, but
its GTID event was saying that it had X+Y bytes.

To fix the above issue, when binlog compression replaces the binlog
cache, it also resets the cache event counter.

RB: 23774