BUG#30578217 : ALTER USER IF EXISTS ...                WITH_MAX_USER_CONNECTIONS 9999 NOT APPLIED                CORRECTLY
Description: When the value of MAX_USER_CONNECTIONS is
modified by the user, the changes are not reflected
immediately.

Analysis: It is seen that when ALTER USER .... WITH
MAX_USER_CONNECTIONS ..; statement is used, the changes are
not applied immediately. The changes are only applied after
all the sessions of that particular user account are
disconnected. So the user_resources value should be reset
every time we change the MAX_USER_CONNECTIONS value.

Fix: The function reset_mqh() is called in
mysql_alter_user() to update the variable user_resources
with the new value.

RB#23653