Bug#30789293 NDB TABLES UNUSABLE AFTER UPGRADE TO SERVER WITH NEW SDI VERSION [noclose]
Problem:
--------

When a MySQL Server is upgraded to a version with a new SDI version, all the
NDB tables become inaccessible. During an upgrade, the synchronization of
NDB tables relies on deserializing the SDI packed into the NDB Dictionary.
This SDI may possibly be an older version and the deserialization is currently
designed to fail should its version not be equal to the current version. This
results in the inability to create a table object in the DD.

Fix:
----

As a short-term workaround, check if the SDI packed into the NDB Dictionary is
of the current version. Rewrite the SDI to have the current version in such
cases to circumvent the check during deserialization.

Change-Id: I7db14c9bf9217db1b14ba079db074a87d2e62cfc