Bug #30850734 ROUTER LOGS WARNING 'UNKNOWN NAMESPACE XPLUGIN' WHEN GR NOTIFICATIONS ARE ON
If you run the MySQLRouter with the use_gr_notifications=1
every few hours it will log the following error:
020-02-03 14:38:46 metadata_cache WARNING [7f10e7edd700] Failed sending ping
on connection to 127.0.0.1:50010; (err_code=5162; err_msg='Unknown namespace
xplugin')

This is because the Router uses obsolete namespace name "xprotocol" when sending
the PING message over the xprotocol connection to avoid the idle connection
being closed. This namespace was removed in 8.0.19

This patch fixes the code to use the correct namespace.

RB: 23781