Bug#30728644 McfErrc's make_error_code() shouldn't be in std::
Problem
-------

Implementing a new std::error_code requires:

- a enum of error-codes
- a static std::error_category
- a is_error_code_enum() for the new error-code enum in the
  std:: namespace
- a make_error_code() that binds error_code to its category.

Only "is_error_code_enum()" should be placed in the std-namespace,
while make_error_code() must be placed in the error-codes namespace
and rely on ADL[1] to locate the proper 'make_error_code()'.

[1]: https://en.cppreference.com/w/cpp/language/adl

In the case of McfErrc and HttpAuthErrc the make_error_code()
was wrongly placed in the 'std::' namespace.

Change
------

- moved make_error_code() out of the std:: namespace
- rely on ADL

RB: 23633