Bug#29527115 MYSQL USES COMPOSITE HASH INDEX WHEN NOT POSSIBLE AND RETURNS WRONG RESULT.
Incorrect number of rows is returned for the partial hash index
while collecting statistics. As a result the check for
"no_valid_range_for_this_index" is passed in get_key_scans_params()
function.

The fix is to return HA_POS_ERROR if the range does not use all key
parts and the index cannot use partial key searches.

Reviewed-by: Chaithra Gopalareddy <chaithra.gopalareddy@oracle.com>