Bug#30789293 NDB TABLES UNUSABLE AFTER UPGRADE TO SERVER WITH NEW SDI VERSION [noclose]
Problem:
--------

The deserialization of SDIs also checks DD version in addition to SDI
versions as detailed in an earlier patch. The deserialization fails
if the SDI's DD version does not match the server's DD version.

Fix:
----

Circumvent the failure by adding a check for DD version similar to
the one added for SDI version previously

Change-Id: I20cccefd4a733d8e6243e9f9b5451300ef8c3f73