Bug#30783011 ASAN: HEAP-BUFFER-OVERFLOW IN MEMCMP FROM GET_BEST_SKIP_SCAN.
Problem:
For a table having a primary key and a secondary key on the same column,
but for different lengths, range optimizer picks the wrong key part in
the secondary index to compare range values. This results in server exit.

Analysis:
For a query having a predicate, range optimizer checks if the secondary
index could be used to do a skip scan. While doing the same, it retrieves
range tree (all the predicates) for columns in the index and performs a
set of checks. The first column passes these checks; so it continues
further with the next column in the secondary index (in this case, its
the primary key column present in secondary index as an extension) to
retrieve the range tree. Since the primary key is defined on the same
column, while fetching the range tree for the keypart, comparison to the
first keypart  in the index (secondary index column) succeeds. This results
in fetching the wrong range tree and when further checks are performed,
server exits trying to compare range values with wrong length of the columns.

Solution:
Use the key part numbers in the index to compare instead of columns while
retrieving the range tree for the indexed column.

Reviewed-by: Chaithra Gopalareddy <chaithra.gopalareddy@oracle.com>