Bug#30792727 ERRORS WHEN BUILDING HASH TABLE FOR HASH JOIN NOT PROPAGATED
When building the hash table for hash join, we did not check for
errors after extracting the value for the hash table key. If the
join condition contained an expression on the build side of the
condition where evaluating the expression would raise an error,
the error did not propagate properly, potentially causing strange
errors later in the execution.

Change-Id: If46fdb0955ce641fc6b9b3369f344a18b82071f2