Bug #30739068 GR NOTIFICATIONS CONFIG DOES NOT UPDATE THE ROUTERS METADATA CACHE
If the user configured use of the GR notifications and some high TTL
then doing shell operations like addInstance or removeInstance could
lead the new change being seen by the Router only after the TTL has expired.
This is caused by the order of the events that occur in such case
1) the change is visible in the GR metadata
2) GR notification is triggered
3) Router reads the metadata (sees it is not consistent) so does not apply the change
4) only after that the change appears in the cluster metadadata

This patch changes the Router behaviour after the 3)
When it notices the incosistent metadata it decreases the TTL to 1s (as it
did when the primary node appeared to be unrachable so far) until
the metadata becomes consistent again.

RB: 23819