BUG#30889192 LOCK_ORDER TOOL MISC ISSUES
Cumulated fixes for known issues affecting the LOCK_ORDER tool to date.

1)

When the file lock_order_dependencies.txt is empty,
the tool reports a syntax error.

Fixed by allowing an empty graph in the debug_lo_parser.yy bison grammar.

2)

The script mysql-test-run fails with --lock-order=1,
and works with --lock-order.

Fixed by using "lock-order=i" when parsing arguments in mysql-test-run.pl

3)

The tool can crash when iterating on the global list of threads.

One issue was that lo_delete_thread() destroyed a thread,
but forgot to remove it from the global thread list,
leaving a dangling pointer in it.

The fix is to add the missing call to
LO_thread::g_threads.remove(lo) in lo_delete_thread().

Another issue was that code iterating on the global thread list
is not always protected by the serialize mutex.

Fixed lo_get_thread_internal_id() and LO_cleanup()
to properly lock the mutex, for safety.

Approved by: Chris Powers <chris.powers@oracle.com>