Bug #30771233: MYSQL_REAL_CONNECT_NONBLOCKING() IS NOT NONBLOCKING RB#23860
mysql_real_connect_nonblocking() had a busy loop in it that was
basically turning it into a blocking function.
Removed the busy loop.
Added a new state for connect to track between connect attempted and
connect complete.
Fixed the test cases that were assuming that
mysql_real_connect_nonblocking was:
1. never returning NET_ASYNC_NOT_READY
2. when it was returning that there are always data to be waited on the
socket. This was dead code test that is now alive.

No test suite added since the existing suite already covers this.