Bug#30273728 SECONDARY LOAD CRASH IN DEBUG MODE
Issue
=====
The parallel reader adaptor used for secondary engine offload needs to convert
records from InnoDB format to MySQL format and hence uses the function
row_sel_store_mysql_rec. This function makes use of prebuilt->index in some
cases which means that the parallel reader code needs to provide the right
prebuilt to the function in case of partitions. This is not possible currently
because it's possible that the parallel reader threads could be working on
records from different partitions parallely and currently we maintain only one
prebuilt.

Fix
===
The correct fix would be to split the prebuilt member, which is currently an
overkill as it holds more information than what it's supposed to, and store the
relevant required information in the parallel reader code pertaining to each of
the partitions. But this is a huge task and requires a worklog to fix it.
Currently as a work around we pass the prebuilt->index information to the
callee as one of the parameters and this prevents the usage of prebuilt->index
in the mentioned function. This fixes the issue because the parallel reader
holds the index information of the tree it's working on as part of the context
and since prebuilt->index would be same as this index we pass this to the
row_sel_store_mysql_rec function.

RB: 23607
Reviewed-by: Mayank Prasad <mayank.prasad@oracle.com>
Reviewed-by: Sunny Bains <sunny.bains@oracle.com>