WL#6057 Make semijoin work with single-table UPDATE/DELETE [patch 2/9, bug with partitioning]
Bug#30787444 MULTI-TABLE UPDATE DOING WRONG ROW UPDATE ON PARTITIONED TABLE

If a table is partitioned by a column, and a multi-table UPDATE
updates this column using some scan, the scan finds a row in partition 1,
the update causes a move of the row from partition 1 to partition 2,
and then the scan continues scanning in partition 2 and finds the row again,
leading to a double update.

Fix: make safe_update_on_fly() realize it can't update on the fly
if a row move is possible, i.e. the partitioning column is updated
and there's more than one partition to update; a temporary table
will be used in this case.
In the test partition_datatype, without the fix, the rows having "the maximum
possible timestamp value minus one second" would become the maximum, and then
wrap around to zero, so after a multi-table UPDATE, we had 4 rows with
value 0, while after a single-table UPDATE we would have only 2.
The fix is thus to use the same detection logic as in single-table UPDATE.
We even improve that logic, for both types of UPDATE, by making it
conditional on updating more than one partition. This way, there are
more chances to avoid the use of a temporary table.
The effect is visible in result files of tests partition_pruning,
partition_explicit_prune, partition_locking in the main suite and
RAPID suite.

Change-Id: I47742630b0a85e92c6c7535aeb06eecb9a346aae