Bug #29488350 HUGE PERFORMANCE DROP ON SPATIAL INDEX AFTER MIGRATION FROM 5.7 TO 8.0
When support for geographical/ellipsoidal geometries was added to MySQL
a performance drop was introduced because mbr_intersect_cmp lost all
specificity, reducing a spatial index lookup to very inefficient walk
through all values in the table.

To fix this mbr_interect_cmp has been reworked to use
boost::geometry::intersects to give the correct result.

Since this bug has been in production for a while a short analysis had
to be done to find what effects this could have on indices that have
used this faulty function. Of the three primary operations on the R-tree
only search uses intersects according to testing. This means that
indices on disk should be unaffected by this bug.

Change-Id: Ia8b4130ee0f07d3f268adfb1301e26d9de2bcc10