Bug #30927590: IMPLICIT CONVERSION FROM 'UNSIGNED LONG LONG' TO 'DOUBLE' CHANGES VALUE
When checking if a number of yottabytes was too large to print,
we would check by comparing the value (in double) to ULLONG_MAX,
which isn't exactly representable as a double. This would cause
the double value immediately above ULLONG_MAX yottabytes to erroneously
pass the test and print as 0Y. Clang 10 correctly warns about
this conversion.

Fix by using the lrint() function instead, which can tell you whether
the conversion overflowed or not. This means we'll need to set the limit
at LLONG_MAX instead, but that should be fine. Update the unit tests
to verify these edge cases.

Change-Id: I46c6a30781a9c5c3c2dc6a32187102afa0ff316e