BUG#30774692 - MYSQL CRASHES WITH SLAVE_COMPRESSED_PROTOCOL=ON
An assert is hit on the replication slave when master is restarted
and the replication uses compression. This manifests in server dying
on the release builds.
The root cause is the compression related context is not restored
when mysql_connect fails. The patch fixes by restoring the replication
context prior to the mysql_connect() failure.

Change-Id: Ia8eaa9aebd1631ba675061ea39c0713a2a22adb7