BUG#30942899 ATRT DOESN'T RESTART PROCESSES ON SOME FAILURES
Fix ATRT to restart cluster processes on all failures. Currently, some
kinds of failures were not being handled correctly so the cluster
processes were not being restarted, leading to issues when running the
following tests.

The investigation of this failure identified three main causes for this
behaviour:

* The code to determine if processes should be stopped was being run
before gathering of results. However, this code had as side-effect
changing the test result which, in case of failure, would lead
restarting the processes. Thus, code to gather results was moved into
the retry loop and will be always executed. This could lead to issues if
result directories from previous tries were left on the environment.
Analysing the code, there was an execution path in which this could
happen, so code was added to remove the results directory.

* The previous refactoring kept the test initialization code but this
caused a change in its semantics. In case of failure inside the retry
loop, those calls would use a 'continue' statement to skip the rest of
code and trigger a new try. Code to restart the cluster processes was
only run after the loop execution was complete. Since now this code is
inside the retry loop, this was causing it to be skipped in the case of
failures. To fix this, the initialization code was refactored to a new
function using explicit returns of status codes, and the control logic
was made explicit.

* In method startAllProcesses, if there was a failure when running
startClusters, it would try to cleanup the status by stopping all
processes since some might be already running. However, the internal
status was STOPPED which caused the call to stopAllProcesses to
trivially return success since there was nothing to do. To fix this,
when there is a failure to start some processes, the status is set to
ERROR, since it is neither running or stopped. This way,
stopAllProcesses will take actions to stop them.

Besides this some other changes were made:
* Refactoring run_test_case and create three functions with the
initialization, main execution loop and result gathering functionality.
This helps making the execution logic more explicit and easier to debug.
* More debug information was added to make it easier to identify
potential issues from the logs. In particular, we now explicitly
identify all conditions which may lead to processes being stopped.
Having this information is important since in the occurrence of a
failure we must be sure that the conditions is set to true even if
restart was triggered by other condition (e.g. a forced restart by
configuration).
- When computing the condition to stop the cluster processes, a skipped
test was being considered as an error. This was fixed so skipped tests
do not trigger cluster processes restarts.

These changes were tested in the following scenarios:
* Run without failures;
* Injecting a failure (killing a cluster process in one of the hosts)
during the test initialization;
* Injecting a failure (killing a cluster process in one of the hosts)
during test execution;
* Injecting a failure (killing rsync in one of the hosts) during result
gathering;
* Changing the assertion in a test to check the behaviour in the case of
a test failure.

Change-Id: Ia6beafae7822c8ef325a1d2b6d9852fae6e83d39