WL#6057 Make semijoin work with single-table UPDATE/DELETE [patch 5/9, bug with different charsets]
Bug#30961924 AND OF TWO PREDICATES WITH DIFFERENT COLLATION ON SAME COLUMN GIVES WRONG RESULT

In this case:
 SELECT ...
 WHERE firstname = 'john' COLLATE utf8mb4_0900_ai_ci AND
       firstname = 'john' COLLATE utf8mb4_0900_as_cs;
doing constant propagation would lead to:
 SELECT ...
 WHERE firstname = 'john' COLLATE utf8mb4_0900_ai_ci AND
 'john' = 'john' COLLATE utf8mb4_0900_as_cs;
and thus, a row where firstname is 'John' would wrongly pass.
The code already had logic for when the collation is my_charset_bin;
here it is improved to handle any collation: if a comparator
uses collation X, and one of the argument is in collation Y<>X,
disable constant propagation in that argument.

Change-Id: I7c1a59639d2463bc4734042719e1ff3352ba3041