WL#13543 InnoDB: Provide partitioning information to RAPID during load
To support partitioned table in change propagation secondary engine requires
InnoDB to provide partitioning information during table offload.

As part of the offload process the parallel read threads spawned by InnoDB
converts the record from the InnoDB to MySQL format and stacks up the records
in a buffer and sends the buffer to RAPID via a callback.

As part of this worklog,

a) the callback provides the partition ID information of the partition it is
working with
b) all records that are present in the buffer sent via the callback will always
belong to the same partition

RB: 23739
Reviewed-by: Mayank Prasad <mayank.prasad@oracle.com>
Reviewed-by: Sunny Bains <Sunny.Bains@oracle.com>