WL#7636: Post-push fix: Add SPJ-API error check for unsupported outer join queries.
When setting up an outer joined scan query, we need the firstUpper() or
firstInner() of the outer joined index scan to be specified.
(The 'first-in-embedding-join-nest')

Furthermore there is a requirement that this firstUpper/Inner is either
one of the ancestor nodes (in the query tree)  of this QueryOperation,
or a sibling of it.

Patch introduce error check and new error codes for this.

Reviewed by: Mikael Ronstrom <mikael.ronstrom@oracle.com>