Bug #29820966 ERROR WHEN FORCING A GROUP RECONFIGURATION AFTER LOSS OF MAJORITY
Problem
========================================================================
On a replication group with 4 members, half of the members are
disconnected from the network.
The other half enters ERROR state due to loss of majority.
When the members are connected to the network again they enter state
ONLINE and see the other half as UNREACHABLE. It should be possible to
reconfigure the group to contain only the members which are ONLINE. On
the test scenario shown below the reconfiguration fails.

1. Setup a replication group with 4 members (M1,M2, M3 and M3).
2. Set variable 'group_replication_unreachable_majority_timeout' to '5'
   on M3 and M4.
3. Disconnect M1 from the network.
4. Disconnect M2 from the network.
5. Wait until both M3 and M4 enter state ERROR.
6. Connect M1 and M2 to the network.
7. Wait until both M1 and M2 become ONLINE. At this point the logs of M1
   and M2 shows the expected error:

       [ERROR] [MY-011495] [Repl] Plugin group_replication reported:
       'This server is not able to reach a majority of members in the
       group. This server will now block all updates. The server will
       remain blocked until contact with the majority is restored. It is
       possible to use group_replication_force_members to force a new
       group membership.'

   After this step the execution of the query...

       SELECT * from performance_schema.replication_group_members;

   ...at M1 and M2 returns that M1 and M2 are ONLINE and M3 and M4 are
   UNREACHABLE.
8. Issue...

       SET @@GLOBAL.group_replication_force_members='M1:port1,M2:port2';

   ...at M1.
   At this point the log of M1 shows the error:

       Plugin group_replication reported: 'Timeout on wait for view
       after setting group_replication_force_members value
       'node1:6606,node2:6606' into group communication interfaces'

   Since both M1 and M2 are ONLINE one expects the reconfiguration to
   succeed.
9. Issue again...
       SET @@GLOBAL.group_replication_force_members = \
           'node1:6606,node2:6606';
   ...at M1.
   Log of M1 shows the error:

       2019-05-23T09:20:38.094222Z 0 [ERROR] [MY-011735] [Repl] Plugin
       group_replication reported: '[GCS] node2:6606 is suspected to be
       failed.Only alive members in the current configuration should be
       present in a forced configuration list'
       2019-05-23T09:20:38.094541Z 25 [ERROR] [MY-011735] [Repl] Plugin
       group_replication reported: '[GCS] Error reconfiguring group.'
       2019-05-23T09:20:38.094610Z 25 [ERROR] [MY-011721] [Repl] Plugin
       group_replication reported: 'Error setting
       group_replication_force_members value 'node1:6606,node2:6606' on
       group communication interfaces'

   This error is also not expected.

Analysis
========================================================================
Hypothesis
----------
The hypothesis is that node2 tries to expel node1 during the time node1
and node2 are disconnected from the network. When node1 uses
gr_force_members, it forces the entire consensus pipeline, which
contains the reconfiguration command where node2 removes node1 from the
membership. We can see this reconfiguration on node1's (and node2's)
error log:

    2019-08-06T08:29:24.596375Z 24 [Note] [MY-011722] [Repl] Plugin
    group_replication reported: 'The group_replication_force_members
    value 'node1:6606,node2:6606' was set in the group communication
    interfaces'
    2019-08-06T08:29:25.046166Z 0 [Note] [MY-011735] [Repl] Plugin
    group_replication reported: '[GCS] Re-using server node 0 host
    node2'
    2019-08-06T08:29:25.046283Z 0 [Note] [MY-011735] [Repl] Plugin
    group_replication reported: '[GCS] Re-using server node 1 host
    node3'
    2019-08-06T08:29:25.046335Z 0 [Note] [MY-011735] [Repl] Plugin
    group_replication reported: '[GCS] Re-using server node 2 host
    node4'
    2019-08-06T08:29:25.046413Z 0 [Note] [MY-011735] [Repl] Plugin
    group_replication reported: '[GCS] Installed site
    start={4317e324 148 0} boot_key={4317e324 137 1} event_horizon=10
    node 4294967295'
    2019-08-06T08:30:24.596689Z 24 [ERROR] [MY-011723] [Repl] Plugin
    group_replication reported: 'Timeout on wait for view after setting
    group_replication_force_members value 'node1:6606,node2:6606' into
    group communication interfaces'

As we can see, a new membership without node1 is installed by XCom, and
we get stuck here.

Evidence
--------
Set group_replication_member_expel_timeout=3600 on the servers.
Run the same experiment. Everything works as expected because there is
no intermediate membership reconfiguration before the one used by
gr_force_members.

Solution
========================================================================
The general idea of the solution is to ensure that forced intermediate
group configurations are ignored. By "forced intermediate group
configurations" we mean group reconfigurations or views that are a
consequence of commands, such as remove_node, that were forcefully
delivered due to gr_force_members.
The actual implementation is twofold.

1. XCom
------------------------------------------------------------------------
Create a new protocol version associated with the following new
behaviour:

* Do not act on forced configuration commands, except force_config.
* Do not deliver forced view_msg.

An XCom instance will only behave like this if the entire group supports
the new protocol version. In other words, either all members have this
patch and behave like this, or all members behave like before this
patch.

2. GCS
------------------------------------------------------------------------
Ensure that the suspicions manager only acts on suspicions when it
thinks the member has a majority at that moment.

RB: 22913