BUG#30947735: NODES ARE CRASHING DURING DISK DATA OBJECT CREATION
Approved-by: Frazer Clement <frazer.clement@oracle.com>

We place the extent pages into the list of SL_LOCKED pages before
the data file is ready and before the mutexes are initialised.
This is not good since we crash on trying to lock extent pages.
These pages are not ready for checkpointing, so we can simply
skip them from checkpointing.