Bug#30908328 - Post-Fix
The MTR test meb.innodb_undo_directory uses a circular reference as an
innodb-undo-directory. An ADD DATAFILE path is not allowed to contain circular
references like "dir1/../dir2".  Instead of changing the testcase, this changes
the code to use the full path of the undo directory instead of the circular path.

Approved in RB#24004 by Rahul