Bug#31024850 data-race on while waiting for shutdown on connection-container
Problem
-------

While the MySQLRouting::start() waits for the connections to close
after a shutdown is triggered, it waits on

    removed_cond

and checks if the connection containers are already empty. It holds
the removed_cond_mutex at that time.

At the same time the connections shutdown and remove themselves
from the container without holding that mutex.

Reported by thread sanitizer (TSAN)

Change
------

- lock removed_cond_mutex while removing connections from container

RB: 24067