Bug #31023004 	THE MTR SUITE INNODB_STRESS IS NOT WORKING ON WINDOWS
Introduced small changes to innodb_stress MTR and python scripts to make the test scripts work on Windows:
The tmpdir path in MTR is required to be an absolute path, so prepending the paths with / is not needed.
MTR's mkdir and rmdir are used instead of unix commands.
SIGTERM is used if SIGKILL is not available (i.e. on Windows).

RB#24062